deleteTodo 함수 수정
* status code를 통해 오류 판별

* 오류가 발생하면 alert창 띄우도록

+++ frontend/src/modules/deleteTodo.js
@@ -5,7 +5,8 @@ export default async function deleteTodo(e) {
   const deleteItem = e.target.closest('li');
   const id = deleteItem.id;
   try {
-    await deleteFetchManager(`/api/todo/${id}`);
+    const result = await deleteFetchManager(`/api/todo/${id}`);
+    if (result.status !== 200) return alert('다시 시도해주세요');
     deleteItem.remove();
   } catch (e) {
     alert('다시 시도해주세요');
